core-mwc: cast args to rol32/ror32 to uint32_t to fix gcc-15.1.1 build warning

+++ core-mwc.c
@@ -163,8 +163,8 @@ void stress_mwc_reseed(void)
 		mwc.w ^= shim_rol32n(mwc.w, stress_get_phys_mem_size() >> 22);
 		mwc.z ^= stress_get_filesystem_size();
 		mwc.z ^= stress_get_kernel_release();
-		mwc.w ^= shim_rol32n(stress_get_ticks_per_second(), 3);
-		mwc.z ^= shim_ror32n(stress_get_processors_online(), 17);
+		mwc.w ^= shim_rol32n((uint32_t)stress_get_ticks_per_second(), 3);
+		mwc.z ^= shim_ror32n((uint32_t)stress_get_processors_online(), 17);
 
 		mwc.z ^= (uint32_t)(id & 0xffffffffULL);
 		mwc.w ^= (uint32_t)((id >> 32) & 0xffffffffULL);
